Greatly simplify title screen data handling
